Simplify loan list with arrays in status type
